pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "trading-env"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "openenv-core>=0.1.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "ta>=0.11.0",
    "yfinance>=0.2.28",
    "openai",
    "python-dotenv",
    "gradio",
]

[project.scripts]
server = "server.app:main"
